requirements.txt
asgiref==3.5.2
Brotli==1.0.9
dj-database-url==1.0.0
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
